requirements.txt
fastapi==0.110.0
pydantic-settings==2.2.1
pyjwt[crypto]==2.8.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.28
